requirements.txt
# 📁 requirements.txt -----

#flask>=2.0.3
#python-dotenv>=0.19.2
#authlib>=1.0
#requests>=2.27.1
Authlib==1.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0